tiktok-platform-microservices/services/iot-worker/requirements.txt
paho-mqtt==1.6.1
redis==5.0.1
pyyaml==6.0.1
python-dotenv==1.0.0
requests==2.31.0
asyncio==3.4.3
